requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6 # Apache-2.0
# Horizon Core Requirements
Babel>=2.3.4 # BSD
Django<1.9,>=1.8 # BSD
django-compressor>=2.0 # MIT
django-openstack-auth>=2.3.0 # Apache-2.0
iso8601>=0.1.11 # MIT
python-mistralclient>=2.0.0 # Apache-2.0
PyYAML>=3.1.0 # MIT
django-chartjs

